Modify src/Day2.hs
Forgot to remove `common`.

The website didn't take my answer the first time so when I was coming
over my code I first tried replacing `intersect` with my own
implementation, `common` to try and be sure... and I forgot to take it
out after realizing the website has a bug.